refactor[#15] : Profile 이미지 컴포넌트 shared 레이어에서 entities/user 레이어로 이동
